Add openSUSE support
While openSUSE is also an RPM distro it is not completely compatible with the RedHat distros. So I had to make this spec file based on the RH one in order for it to work on openSUSE and OBS. 

Some changes: 
 * Uses /usr/lib/systemd/system to store the service because SUSE doesn't allow /etc/systemd/system to be used by packages
 * The package for go is called `go` on openSUSE and not `golang`
 * Change license name to the one SUSE allows. You should use identifier in https://spdx.org/licenses/ 
 * Because OBS build environment has no internet connectivity, you have to add another source with all of go's dependencies and make it the $GOPATH
 * I changed the file mask for systemd service from 0755 to 0644 

I don't know how to solve the warnings RPMLINT gave me but they don't seem to cause any issues with the package. 

The openSUSE repo I created is @ https://build.opensuse.org/package/show/home:ranyy/yggdrasil